Replaced getUserLoggedIn() C++ function with more robust version, as documented by MS here: http://msdn2.microsoft.com/en-us/library/aa375400.aspx
Modified vcproj and VC files to update them to newest 2008 version of Visual Studio.  Also added the Secur32.lib to the linker's command-line options (needed for user logged in functions).

Added getUserLoggedIn() as a function of the ISystemInfoProvider interface.

Added a stub function for the Linux class ProcFileSystemProvider.

Added proper provider call to SystemInfoUtils.