Sanitize username/password acl tokens before allow client to retrieve with getAuthz() resource property calls
Also added methods to all AclEntry types to allow for the future sanitization of credentials, and added appropriate method to the IAuthZProvider interface

Default behavior if you do not call method with sanitize explicitly off is to sanitize, this may not be desired if it is an internal container call (like QueueSecurity), as the full token is necessary to call a check permissions.  Check permissions on username/password now will fail if the password is empty or null (to prevent sanitized password causing security hole and it is just not a good idea to accept empty password).