Archiving old version of DLog project.
This code was developed on the GAML trunk from pre-SAML conversion and the code was well out of date. Changes have been transfered to a new project based on the SAML trunk, so this code is obsolete. This commit serves as a final archive version for future reference if needed, and should not be used in general practice.