Last step of catch-up since base revision change, I think.
This one changes everything to use a separate LoggingContext instead of piggy-backing off the Calling/Working context types. This simplifies and cleans up pretty much all of the implementation, so we don't have to worry about how the security is handled or when a thread has no working context, etc. As a result, the majority of the changes to Calling/Working context and the classes that handle them have been rolled back.

Introduced a logging context variable on most of the threaded workers. May be some still without. Focused mostly on RNS and BES workings, to ensure correct hierarchy. GUI and ResourceFork interfaces, and CLI all seem to be working correctly, with one snag noticed when multiple Logger resources exist on the same container but don't use the same database.

Some changes also to data structures in the Queue code and associated Queue database structures to help maintain logging context during job execution. Similarly with Persistent Outcalls. These should not effect any external resource communication, they are strictly internal representation changes, but this means that this code in unlikely to work with existing state directories (mostly untested, but some evidence to support this hypothesis was observed).

**Due to the severity of these changes, it's recommended that prior revisions not be used, unless this one proves to have other, disastrous results. Keeping the existing Context types unchanged is greatly preferred.**