Checking in as a check-point. Don't use it.
Reverted TransientCredentials to what it already was (in theory, I didn't actually revert...), and switched the ICLogin tool to use a TrustCredential instead of my own derivative type. The type is accepted, the private key signs it (using bouncyCastle stuff), but the STS won't grant access based on it. Other port types seem to be granting to the CILogon cert, though, if created after obtaining the CILogon cert. So there is probably some part of the cert that differs between logins that I need to pinpoint, and then base the access control on only the part that's constant (I feel like Andrew and Jim talked about this at some point, but it all pretty much went over my head; maybe I can talk to him or Chris tomorrow about the pattern stuff again).