fix for JobApplicationContext and helper_methods
in JobApplicationContext, a remove was being called erroneously on the entire collection being processed rather than on the current item to be processed.
helper methods had an undesired '=' sign on the end of an unset statement.
GenesisIIBase just has a formatting change, to indent some code more carefully for when the new (and commented code) around it is turned back on.