updates to improve output for clientserver
some errors were not printed to the client, but should be now; the exception handler manager's EH function was saying things we want to send to the client, and now it uses the right outwriter.  cleaned up actual exception handlers a little also, to simplify number of them and ensure we always say something to the client about what went wrong if possible.