Updated to support two things.
First, every second processwrapper will do a getrusage and save the results
to the rusage.xml file. It does this rather than wait for termination
because if the job is killed by SLURM, PBS, etc for exceeding
its wall clock limit, we have no data on how long it ran. Similarly
if the node fails, we do not know anything about exection time.

Second, on SIGTERM it also dumps the data.